# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "API description"
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "Returns a hello message",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/generate/{name}": {
      "get": {
        "description": "Generates a JWT token for the given user",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "User name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/protected": {
      "get": {
        "description": "Returns a welcome message for authenticated users",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/sample/{method}/{number}": {
      "get": {
        "description": "Returns information about a sample",
        "parameters": [
          {
            "name": "method",
            "in": "path",
            "description": "Sample method",
            "required": true,
            "type": "string"
          },
          {
            "name": "number",
            "in": "path",
            "description": "Sample number",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "info": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}